Fix font misalign turning font.png off on LÖVE 0.10
To do this, when we exit Ved options and handle calling
`temp_print_override()`, we also handle calling
`undo_temp_print_override()` if we're not going to call the former. The
problem was that we were letting the effects of `temp_print_override()`
stand when we shouldn't be doing that.