Add clarification to temp_print_override()
The fact that all the names have '11' in them make it seem like that
this function is only used for LÖVE 11 and upwards, but this function is
also used for LÖVE 0.10 as well, so the names are a bit misleading. I've
added a comment to clarify so we don't misunderstand in the future.