Enable screensaver
It's always annoyed me a bit how Ved will stop my screensaver whenever
it's running, and prevent the computer from automatically locking up
from inactivity.

It's always done this because Ved runs on LÖVE, which itself runs on
SDL. By default, SDL has disabled the screensaver since version 2.0.2
(and LÖVE is guaranteed to use SDL 2 since 0.9.x and onwards), since it
is primarily a game-making tool, and it would be rather inconvenient for
your game to NOT disable the screensaver while it's running.

However, since Ved is more like an editor rather than a game itself
(unless you count Vis as a """game"""), we will want to re-enable the
screensaver again.