Improve assets Reload button yellow flash
It now uses the generic timer, like "Copy script" does. Instead of doing
the hack of drawing a yellow button on top and then calling
`love.graphics.present()` immediately (which can break easily if it's
done somewhere else (e.g. an abstracted function in `func.lua`) and
we're on a non-1x scale). So now it will be yellow for more than 1
frame, and is less prone to breakage.