Fix state 15 s.smallerscreen overlapping issues
State 15 is the common state for the help articles, level notes, plugins
list, and update info. By the way, we really need a common name for
state 15 that isn't "help articles/level notes/plugins list/update
info".

The issues were that it visually looked like Ved was accounting for the
fact that the article list on the left would cut into the content on the
right, and so had this toggle where you could choose which one would
have priority - however, it was only visual priority. You could still
click on the buttons to the articles behind them while the article
content had priority, and the article list would get scrolling priority
when it visually looked like it shouldn't.

So now that's fixed. But not without my specialty - an ugly hack!

Regardless of the collision issues, there was still a visual bug where
the buttons of the article list would still highlight if your mouse was
on top of them, even if the content had visual priority. So I had to
basically override the alpha `hoverrectangle()` would use. Without
inlining the code entirely (which would be WET), I decided to go with
the fact that it won't highlight it if `nodialog` is `false`, but I also
didn't want to mess with the `nodialog` variable for too long, so I had
to temporarily store its old version, then set it to `false`, and then
re-set it back to what it was before. It's ugly, but it works.