Cancel editingbounds when going to the map screen
I made sure to make the cancelation undoable if only the top-left corner
has been placed, too.

It seems like Ved didn't intend for you to use the map screen while
editing boundaries. Ved won't let you use the arrow keys to change rooms
while you're editing boundaries, in both the main editor and the map -
but it forgets that you can just go to another room using left-click on
the map, or right-click to go to a room without closing the map.
Clicking the top-left corner in a room and then moving to another room
to place the bottom-right corner will just make the un-properly-placed
corner 0,0 or 320,240. Which is pretty weird and very unintended. Not to
mention the fact that you can undo/redo while on the map screen, and
copy and swap rooms.