Disable script jumping if script to be jumped to is the current script
Script jumping was already disabled if the script to be jumped to was
already in the script history stack, but I forgot to account for if the
script to be jumped to was already the current script. Not that it
matters, because if you do jump to it, you can't jump to it again
anyway, due to the previous "don't jump to a script if it's already in
the stack" rule. But still.