Switch subtool hotkey printing code to tables
Today, I noticed that there was a weird glyph next to the custom brush
size subtool, even though that glyph doesn't look like any pressable
key, and also there's no hotkey for the custom brush size subtool. I
swear I didn't do anything to make that happen!

Well, I did. `git bisect` traced this back to
177d0f9e81ccd637c79d8b7fee873e4538ddd580, the addition of the Space
glyph into tinynumbers. It turns out that Ved was printing a Space
because that's basically the hotkey specified to be rendered for that
tool, except that until I added a Space glyph to tinynumbers, there WAS
no space glyph, so it just didn't render anything, as intended. Also
there wasn't enough room to print the full space character, so it just
rendered the first few columns of pixels of it, resulting in a weird
glyph.

It had to put a placeholder in the form of a space because it was doing
`(" ZXCVHB F"):sub(k,k)`. This way, you can't make the 8th character of
a string literally nothing, because then it would pick what you intended
to be the 9th character.

So I did what should've been done in the first place, and did `({"",
"Z", "X", "C", "V", "H", "B", "", "F"})[k]` instead. Apparently this is
called a "rope". Now, you CAN make the 8th index of a table literally
nothing, so this is fixed now.

(Oh, and the glyph didn't render for the first subtool, because
apparently there's a conditional where it only renders for k >= 2 and k
<= 9.)