Fix no-dialog-anims immediate OKing of dialogs canvas-clicking-created
Or, in a less condensed version: Fixed being able to accidentally close
dialogs that were created by clicking on the main canvas if you had
dialog animations disabled.

The most common case of this would be if you tried to add a 21st trinket
or crewmate, but the place where you tried to place it just happened to
be the OK button to the dialog, immediately closing it.