Added fallback font when using font.png
The entirity of the Space Station TTF is now exported to two .pngs in
LÖVE's ImageFont format. These fonts are fallbacks for the regular
font.png when it is used, for a few reasons:
- Unsupported characters were previously not displayed at all, so
  displaying anything ought to be better, even if it's not in the same
  font. Non-ASCII replacements cover most languages, but not Russian,
  and the language picker dropped the ç in Français because it's out of
  scope for replacements (might need to think of special font.png names
  for languages). The non-ASCII auto-replacement feature will still be
  used, as just fully using the same font will probably look better than
  having fat C64 characters in the middle of everything.
- I had a second reason but it's probably somewhere in the previous. I
  guess it had something to do with the places non-ASCII auto-
  replacements do not cover: filepaths, level notes, anything else that
  doesn't come out of language files.

If anyone wants the code that converts TTF to ImageFont .png, just tell
me, I made it as part of a dev plugin.

[pre04]