Correct gravity line hitbox length
The length is actually 2 pixels more than the visual. That is, there is
1 more pixel of hitbox on each end of the gravity line.

You can test this by locking a gravity line and then place a tile on top
of one end, then you'll find you can still interact with it at that end
even though it looks like it's in a tile and should be un-interactible.