Refactored font and language loading
I'm making the changes I mentioned here:
https://gitgud.io/Dav999/ved/merge_requests/30#note_74644

There's now a function loadfonts(), which loads either the TTF font or
font.png, and should handle anything related to that. Previously the TTF
font would be created first, then maybe font.png, and changing the font
back to TTF would cause the TTF fonts to be loaded again, but from a
different place in the code. Now this is all in one place - if the font
setting changes then loadfonts() can be called, that's it for the font.
In the few cases where the fonts are needed before main2 would load
them (crash screen, plugin error for main2, outdated LÖVE), the
loadfonts() function is just called there. Except the "Loading..." text,
which I decided to just turn into an image. That may seem cheesy, but it
avoids loading an entire font only to possibly replace it afterwards
(and splitting font loading across multiple places, avoiding which is
the point of this change) and looking like the font is loaded early when
actually it isn't.
The "Loading..." text is also displayed at double size now.

There's now also unloadlanguage() and loadlanguage(). More or less the
same principle, loadlanguage() is called when starting up, and when the
language is changed. unloadlanguage() was needed to get rid of the
s.plang variable. That function is called, then s.lang is set, and then
loadlanguage() loads the language and may reload files that depend on
them. (Well, I say plural, but const is the only one for now.)
loadlanguage() also handles non-ASCII replacements for font.png.

English is now also not unconditionally loaded before another language.
The idea was it could be a fallback when strings exist in English but
not in the selected language, but nearly all strings are part of a
table that just gets overwritten completely, so it's just a waste of
resources. And if the selected language file doesn't exist, then
English is loaded instead.

Some parts of the loading sequence are now reordered. For example,
loading font.png requires the OS-specific function to read files, and
the OS-specific VVVVVV folder path, which were previously needed much
later than when the fonts were loaded.

Still TODO: loadtinynumbersfont().

[pre07]