Stop highlighting if unfocused
Now that if you click on Ved to bring it back in focus nothing will
happen, it's a bit misleading to show a button being highlighted when
Ved is not focused. That said, I did leave the cursor rendering in case
someone wanted to cross-reference, say, room positions with some
external source of information while Ved is unfocused, and I also made
it so the current tool border gets lightened when Ved is unfocused,
which is a hint that clicking back onto the main canvas while focusing
Ved again won't place a tile or anything.