Don't stop input going to state 15 from state 3
State 15 is the level notes / plugins list / help articles / update
info, which I will now refer to as the 'article viewer'.
State 3 is the script editor.

Turns out I was a bit overzealous with
4372816495f1beae65237601fa5e0516c292e197. Typing in the article viewer
coming from the script editor won't type anything in the script, so it's
unnecessary to stop input. In fact, stopping input here means you'd have
to close and reopen the script again in order to be able to type again,
which is a bit annoying.