Add the start of new input system
WIP.

This is the one to replace the "2 or 3 variables concatenated together"
system of before, as discussed in #86.

Currently, the interface to create and manipulate inputs is finished. To
keep things consistent with `dialog.create`, you use `input.create` to
open a new input buffer, and `input.close` to close it. So this way you
don't have the potential to overlap inputs, since there can be more than
one input at a time. However, the plan is that the user will always
manipulate the last buffer opened, and buffers can be bumped. The plan
is also to tick the blinking cursor timer in `love.update` as before.

I've created a new state, state 34, to test the new input system. Once
you go to it, don't go back or it crashes Ved. One of the things I
didn't expect to run into was the Y misalignment correction - so I had
to uncorrect by moving the blinking caret upwards 3 pixels. This is
currently hardcoded, but I expect I now have to recapitulate the Y
misalignment correction code, only that I'll be doing it in the other
direction for the caret.

Other things that still don't exist yet: blinking caret for
center-aligned text, and support for LÖVE 0.9.x versions since
`Font:getWrap()` doesn't give us the information we need to position the
caret correctly in those versions. The plan is to manipulate text with
`love.textinput` and `love.keypressed` as before.

And of course, as this is just the start, I haven't replaced any of the
old input usages yet.