Add input.leftmost/input.rightmost
These are basically just Home/End for inputs.

One of the things I like about (n)vi(m) is that if you type $, it goes
to the end of the line, and for each line you go to afterwards the
cursor is also at the end of their line, even if the line you go to has
more columns! So I added that behavior to Ved, too.