Add selection visuals in multiline inputs
This is the start of adding support for selection to the new input
system. That is, you hold down Shift and then press any direction key to
start selecting text in that direction.

This only adds support for the visual part of selecting text, and only
for multiline inputs. Currently, you cannot copy, cut, or paste text.
Support for that will be added later.

This renames the `input.drawcaret()` function to `input.drawcas()`, for
"draw caret and selection".

As well, I happened to use `love.graphics.getColor()` to get the current
color so I can set it back after I'm done drawing the selection
rectangles. Apparently, this function isn't hijacked in
`love11compat.lua`, resulting in me setting the color essentially to
black, so now I've hijacked that function so it works properly in LÖVE
11+.