Fix starting selection at boundaries of input
This is so if you are on the leftmost part of the line, for example, and
do Shift+Left, it doesn't start a selection.

The previous code was buggy because it didn't take into account if the
input was rightmost or not. Also because it didn't consistently use
utf8.len() everywhere.