Add support for LÖVE 0.9.1
So what I did was re-implement the C++ code of the utf8 module into Lua.

It was very annoying, because I had to deal with constant off-by-one
errors everywhere, since Lua is 1-indexed and literally every other
language (especially C and C++) is 0-indexed. All that time saved not
having to internalize 0-indexing for newbies has been wasted by people
who switch between programming languages having to deal with off-by-one
errors.

But the annoyances don't end there, because Lua doesn't have native
bitwise operators until Lua 5.3, so I had to import the `bit` module to
use them. Also there was that one part with the fact that in C/C++,
null-terminated strings are a common thing, but in pure Lua you can't
check for the end of a string by checking for '\0' so I had to fix that,
too.

At the end of the day, this was a lot of work just to support one
version.