Add support for LÖVE 0.9.x printf wrapping
Added a new function to func.lua, Font_getWrap9(), which is basically
Font:getWrap() in LÖVE 0.9.x if it returned the table of the wrapped
lines like it started doing in LÖVE 0.10.0.

This also fixes a bug previously where we assumed that if the printf
text was only one line, it would be fine. As it turns out, printf
ignores multiple spaces between words, so it wouldn't have been fine.