Add Ctrl+Left/Right
Oh wait, apparently it's Option+Left/Right on macOS (Option is just a
renamed Alt). Silly Apple! To reflect this, I've made a new variable
with the same purpose as `ctrl`, named `modifier`, which is Alt on macOS
and Ctrl on every other OS. Also, apparently there's a Ctrl+K on macOS
(as opposed to Cmd+K, as currently Ved uses Cmd+U/K on macOS) which is
like Ctrl+K, only it deletes to the end of the paragraph instead of the
line, and there's no Ctrl+U counterpart because Ctrl+U is used for
toggling underlining. I don't feel like it's a good thing to make Cmd+K
Ctrl+K instead, or to do the same thing with Cmd+U when there's even
less of a good reason, so I'm keeping Cmd+U/K as is.

Anyway, currently this goes between words (defined by "things between
spaces"), but it can be made to go between any arbitrary delimiter just
by changing the separator from a space to something else. Like if we
ever include typing directory separators into input fields (which we
aren't doing right now), we can make that separator be, well, the
directory separator. Or more relevantly, we can do this for scripting
commands by making the separator any of '(', ',', or ')'. Although
Ctrl+Left/Right conflicts with using Ctrl+Left/Right to jump forwards
and backwards in the script jumping stack... Hm.