Add Ctrl+Shift+U (hexadecimal input)
Yes, you can now insert <0x20 chars easily. And also Unicode chars if
you really want.

This adds `utf8lib_092fixedchar.lua`, which is yet another
re-implementation of a `utf8` module function, because `utf8.char()`
was broken when it was added in LÖVE 0.9.2, and was only fixed in
0.10.0. So I needed to copy-paste the fixed function but in pure Lua.

The look of the hexadecimal input isn't too great (just a rectangle
with no blinking cursor), but I didn't want to re-position the caret,
and so it'd be bad to have it blink before the hexadecimal input anyway.

The hex input will probably get scissored away in cropped inputs that
can overflow if you're on the right end of the scissor, but ah well. I
think we can just position it such that it starts shifting to the left
earlier but then doesn't get cut away by the scissors.