Clear selection pos on mouse release if applicable
We don't want to reset it while selecting, since the selection pos is
initially the same as the caret pos. But we don't want to leave it
hanging around, since if you type something it'll select it instead of
clearing it. So `love.mousereleased()` is the perfect place to reset
this.