Add grouping together of UNRE insert/delete operations
The way it works is that when you do an operation that is explicitly
inserting a char (including finishing a Ctrl+Shift+U, and including
inserting newlines, and does not include pasting), it gets tagged as
`UNRE.INSERT`. If two `UNRE.INSERT`s come together one after the other,
they simply get grouped together in the same operation. The UNRE
operation gets untagged if the cursor is moved. Same goes for
`UNRE.DELETE`, which applies to deleting chars forward and backward, and
includes delete-concatenating lines.