Rename input table to newinputsys
This is to ease transition, where when the transition is finished, we'll
rename it back to input.

All newinputsys.active conditionals (previously input.active) are also
now guarded against newinputsys being nil, but only because we're in a
transition right now.

This also means once you go to state 34, you can go back without
crashing.