Added start of GUI overhaul
TL;DR: Code quality improvement.

A lot of this is still subject to change, but here's the start of the
system I've mentioned a couple of times and explained the idea of here:
https://gitgud.io/Dav999/ved/merge_requests/31#note_76779

Basically, Ved has mostly been using functions like love.draw and
love.keypressed directly. Some states had drawing functions that were
so long I split them into their own file, but many of them were drawn
in love.draw directly. This overhaul makes it possible to have specific
callback functions for each state, in their own file, in a "uis"
folder. And not just for drawing, but also for update, keypressed,
mousepressed, etc. There's also a "load" function to not need to put
state loading code in the `loadstate` function in func.lua.

Now, the most fun (and code quality improvement) comes from the concept
of Elements. Right now I think one of the nastiest things in Ved code
is that drawing code is used both for drawing buttons (at a hardcoded
position) and then checking whether that button is clicked, while more
and more buttons have gotten hotkeys, which are mostly checked in a
MASSIVE elseif chain in love.keypressed, which is supposed to run the
same code as the drawing function does when the button is clicked. If
you don't want to duplicate this code, then you have to make a special
function just for it, and put that somewhere else.

Now, every state can also have a list of interface Elements, which can
be predefined objects, like buttons, or containers that automatically
position child elements, or other things. Instead of what I described
in the previous paragraph, buttons could just be defined with a label,
a function and a hotkey, and drawing, positioning and checking for both
clicks and the hotkey being pressed is taken care of automatically. I'm
simplifying, but still.

This should also make it easier for plugins to add their stuff to the
interface. If plugins want to add a button in the new system, they can
just add it to the list and won't have to worry about clashing with
where another plugin has put their changes. Though, I'm also thinking
of an even more robust system in addition (a special place in each
state where plugins can add buttons) in case things won't fit on screen
otherwise.

I started with the map state, which is not totally "converted", but
the "Save map" and "Return" buttons (along with their hotkeys) already
are. There's also more demo/debug stuff in the map, which is why
there's a temporary demo button next to the map, which shouldn't be
much of a problem.

[pre22]