Fixed GUI button labels not changing on language change
Strings in GUI elements (thus far, this only matters in the label of
LabelButtons) are evaluated when the GUIs are loaded during startup,
and changing the language doesn't change those like it does in most
other cases where the indexing on L simply happens whenever a string
needs to be printed.

I don't want to create special "notation" for language strings in GUIs,
like a function that returns a function that does the indexing, so now
all GUIs will be reloaded when the language is changed. Hopefully we
won't want to use state in Elements for anything too important (you
would see scrollbars reset for example)

[pre25]