Added some shortcut replacements in advance of new input system
!31 refactors the input system, and comes with some shortcuts that can
be used in every input. These conflict with a few existing shortcuts in
the script editor and levels list, which are not as important, so these
shortcuts are now being replaced by new ones, while the old ones can
still work until the new input system is actually merged in, and give a
warning to use the new shortcut.

The shortcuts are:
- Ctrl+left / Ctrl+right -> Alt+left / Alt+right (jump into and out of
  scripts in the script editor. Everyone should know this shortcut,
  you're missing out if you don't)
- Ctrl+A -> Ctrl+R (Assets. Mnemonic: Resources)
- Ctrl+D -> Ctrl+F (Open lvl dir. Mnemonic: Open levels Folder)

The function show_notification is now added and causes a "notification"
to be shown in the bottom left corner for 5 seconds. You'll see this
when using the old shortcuts, to get used to the new shortcuts.

[pre27]