Added level sanity check for duplicate script names
Joshua discovered a level had two scripts that had the same name,
which shouldn't be possible. Since it may cause issues in Ved (a crash
when deleting one of the two, and editing either will edit both at the
same time) a sanity check for this when loading the level is a good
idea. It will still overwrite the contents of the first script with the
second, but it will no longer insert the name a second time, and will
raise a warning.

The original cause of the duplicate script remains unknown (a mistake
during manual editing is always a possibility), but thanks to the
people who tried to find an explanation for it!

[pre31]