Add VVVVVV color function(s) and use them
Also, this fixes enemies in the gray Warp Zone tileset not being green.
Complained about in https://tolp.nl/forum/index.php?topic=3934.0 ,
noted in https://gitgud.io/Dav999/ved/issues/21#note_23385 . Space
Station tilecol -1 enemies are also green instead of gray. However,
platforms/conveyors in Warp Zone tilesets are still gray instead of
green.

This adds vvvvvvfunc.lua, which is basically code from VVVVVV, but
rewritten in Lua. It contains partial UtilityClass and Graphics objects,
as well as the functions UtilityClass::updateglow() and
Graphics::setcol(), and partially Graphics::drawentities() and
Graphics::gamerender(), because the gravity-line-related variables
`linestate` and `linedelay` are handled in the former, and the color of
roomtext and roomnames are handled in the latter. Also it contains
fRandom() and RGBf().

Most of the functions were decompiled and re-written by yours truly, but
Graphics::setcol() was actually posted by Terry Cavanagh himself (
http://distractionware.com/forum/index.php?topic=1992.msg21312#msg21312 )
(except for a red and blue swap with color 36).

A new attribute has been added to the tileset information in const.lua,
`v6col`. Which is simply the color given to Graphics::setcol(), the
color function, that the enemies of that tileset use.

The colors updated include:
 - Every entity drawn, including every enemy of every tileset and
   tilecol
 - The middle-click trinket
 - The trinket/crewmate/player locators on the map screen
 - Roomtext and roomnames

Closes topic=3934.0.