Add minimap preview support
Here's how minimaps work:

If your map is within 5 by 5, then you get the 4x-zoomed minimap image
that's 48 by 36. However, even though it contains every single detail of
the room, that's a bigger image than the size of the room in tiles, so
the image is stretched and distorted!

If your map is within 10 by 10, then you get the 2x-zoomed minimap image
that's 24 by 18. This uses a tile grid that doesn't have the neatest of
patterns because it doesn't divide evenly with 40 by 30. It also misses
some tiles on the right edge and bottom edge.

Otherwise, if your map is bigger than 10 by 10, then you get the
1x-zoomed minimap image that's 12 by 9. This uses the neatest
evenly-spaced rectangular tile grid, but it's all concentrated in the
top and middle of the room, so tiles on the left edge, bottom edge, or
right edge simply get missed.

You can now hold down semicolon (";") to see the minimap grid displayed
on the room in question! Although, since I made it be drawn in
`displayroom()`, entities are drawn over it... I'll have to fix that
later.

I've also added a function to preview the minimap image itself. I'm
using it in the bottom-left corner, but it's drawn on top of the fill
bucket because the UI is so cramped I couldn't find a good place to put
it other than that, so I've left it commented out for now. But you can
uncomment it at the bottom of `drawmaineditor.lua` if you want.

Eventually, I want the minimap preview to be visible at all times, and
not hidden behind a keybind, or even worse, a button. Because then it
would get too annoying to make sure your minimap image looks nice. I'd
also like to put something in the map screen that previews the minimap
image of the entire level - there's much more space there, but I haven't
done it yet.