Add playtesting feature with initial support for Linux
It's finally here! Ved now has playtesting! (When this gets merged...)

Here's how it works: all you have to do is press the Play button, which
has replaced the Help button, and the Help button has been moved to
between the Undo/Redo buttons and Cut/Copy/Paste buttons, just like in
Vis.

But anyways, then Ved will let you place Viridian wherever you want, and
they will snap to checkpoints and the start point if you put your mouse
on them. You can also move around the map as well while doing this,
although everything else is disabled, but you can cancel playtesting by
re-pressing the Play button, which is now an X.

Then when you've finally made your choice - just left-click, and you'll
start playtesting right there!

Shoutouts to @leo60228 for making libvloader, which makes all of this
possible.

Now for the technical details...

This commit uses libvloader alpha 5, which I've named to be
libvloader_linA05.so. I didn't want to disrupt the "_linXX" naming
convention by adding another underscore, but I needed some way to
differentiate that it was an alpha... so I added an A.

There is still a lot incomplete with this playtesting feature. For one,
Ved is currently not properly equipped with the right filefuncs that
this feature requires! Ved cannot make symlinks, so I've had to tell the
user to do it themselves. Ved doesn't have good enough file- and
directory- checking functions, so right now playtesting_validate_path()
just always returns true.

And there's this annoying bug where if the symlinks to vvvvvvmusic.vvv
and mmmmmm.vvv are broken (i.e. the actual files don't exist), then
VVVVVV will just not play music at all. It would be nice if Ved could
detect broken symlinks and remove them... but it can't detect broken
symlinks yet.

Also, making the playtesting_askwherestart mode was a bit hacky. For
one, when I darkened the side panels, I only did it AROUND the Map
button, and not, say, not draw the map button, then darken the side
panels, then draw the Map button on top of the darkening. Same thing on
the map screen too, where I didn't entirely cover the part of the map
screen where the roomname is, so theoretically if someone where to have
a long enough roomname, it would abruptly get darkened...

But the main point is, I added playtesting_askwherestart to the nodialog
calculations, and then carved out exceptions for
playtesting_askwherestart in the code, but only if coordsdialog.active
was false. Very hacky. It got so hacky that the Return button doesn't
highlight properly when you hover over it in the Map screen if
playtesting_askwherestart is true, but I decided to not bother at that
point. Hopefully Dav can come up with something more elegant for
playtesting_askwherestart in the new UI system.

New images have been added - play.png, playgray.png, and stop.png.

New glyphs to tinynumbers have been added - 'g' is Viridian
rightside-up, 'v' is the flip symbol, 'j' is Viridian upside-down, and
';' is an unlocked lock. They're used in the bottom-left to signify that
you can flip the player with Shift and prevent snapping to
checkpoints/the start point by holding Alt.

A new config option has been added - s.vvvvvvnonsteam. This is set if
Ved can't locate the Steam version and needs to ask the user where they
put the non-Steam version of VVVVVV.

And lots of devstrings have been added.