Disable minimap grid preview outside 20x20
VVVVVV does not render minimap images for rooms outside the normal 20x20
area. So in those rooms, pressing semicolon (';') to preview the minimap
grid is disabled.

Furthermore, I've added a note to the temporary commented-out minimap
preview that it shouldn't render outside the normal 20x20 area. In the
interest of preventing unnecessary merge conflicts, I've only added a
note, not changed the commented-out code.