Fix renaming references to scripts escaping pattern chars in entities
1a7e96c1e2d0797e360fe1a649ba94a04cc59a2e applied `escapegsub()` to the
old script name, because the code to rename references to scripts needed
to use `<string>:gsub()` on script lines. However, this escaped version
was also being used when it was just directly comparing the `data`
attribute of script boxes and terminals to see if they referenced the
old script name, making this an unnecessary use of `escapegsub()` for
those cases. This has been fixed now.