Add conversion from frames to seconds in side panel
Now if you put your cursor on a line with a command that takes in a
number of frames (these commands: "delay", "walk", "flash", and
"shake"), a helpful string of text will show up in the side panel on the
right that will convert whatever mysterious number you just typed into
its equivalent amount of seconds. E.g. "30 = 1.02 seconds".

I've made it so nothing will show up if the amount of frames in the
command is 0 or is negative.

A devstring has been added for this.