Account for out-of-bounds coords for {show,hide}coordinates
Turns out, they operate on the concatenated rows system - that is, it
takes the X and Y you give, and gets a room number that is just X +
Y*20. So whichever room you're targeting basically wraps around if you
go past in the positive-X direction, but all the rooms are shifted up by
one. Basically reminiscent of how room properties work outside 20x20 in
bigger than 20x20 maps. So `showcoordinates(20,0)` would end up
targeting (0,1).

I've added a new script context "roomnum" for this, in case the room
number isn't in-bounds of any room-related array in the game (which are
all 400 rooms). Then it'll just say "Room <roomnum>", like "Room 400" or
"Room -1". I've also added a new script context "roomnumscript", even
though "roomscript" is currently unused (and so is "roomnumscript").