Show track number and name for music commands
Now if you put your cursor on a line with any of the commands `music`,
`play`, `stopmusic`, or `playremix`, in the side panel it will show you
the track name and number.

For the `music` command, it simply takes as many digits as it can until
the first non-numeric character, and parses that as internal. Else, if
it's all digits, but if it's out-of-bounds of the simplified IDs
(greater than 11), it will just parse it as internal. When parsing as
internal for both `music` and `play`, it modulos it by 16, so `play(20)`
would actually be `play(4)`. But negative numbers don't work at all for
either.