Did a little bit of housekeeping
Mainly, part of the obsolete pieces of commented out code have been
removed all over the place as I saw them, and the quality of some code
has also been improved.

Again, this is only part of it, I didn't structurally go through all
the code, so I only fixed parts I happened to find by scanning through
or searching code.

Also, fixed a bug in the GUI overhaul system where dismissing a dialog
would cause the keypress to be repeated outside of the dialog if dialog
animations were disabled (such as, pressing esc in a dialog in the map
would immediately exit the map screen)

[pre39]