Add integrity checks for null bytes
I've added an integrity check that there are no literal '\0' bytes in
the level file, and I've also added an integrity check that there are no
XML null bytes such as '&#0;' or '&#x0;' in the level file, too.

VVVVVV, being coded with C++ (but using TinyXML to read XML files),
doesn't deal well with null bytes, and especially not literal ones. It
will basically stop reading the level file when it sees a literal null
byte, and it will basically end the XML tag if it sees an XML null byte
and parse it as if it was just suddenly cut off for no reason. Ved,
being coded with Lua, deals somewhat better with it, but it will
abruptly stop printing a string if it sees a null byte. So let's just
not deal with nulls.

This deals with the "can't handle 0x00s" part of #77.