Add platform labels for `p1`s of 4 and negative `p1`s
Just checked just now, and platforms with a negative `p1` seem to all
stay still. (Note: I haven't checked this in the decompiler yet, but I
have no reason to believe any particular negative value will dissent
from the norm).

And platforms with a `p1` of 4 all just move right until they hit a
wall, and don't obey platform bounds. All others just result in a
missing entity (note: I haven't checked this in the decompiler yet, or
even checked with GDB to see if the entity/block is still there, but my
intuition suggests that this is the case for every single `p1` above 8).

So I've put the label "[]" on platforms with negative `p1`, and I've put
the label ">|" on platforms with a `p1` of 4.