Remove duplicate entity when canceling copy
Turns out this was how Ved worked: When you moved an entity, you just
moved an entity. But if you copied an entity, first Ved already makes a
copy of that entity, and then you're actually moving that entity, so
you're really just actually doing an entity move but the entity was
duplicated for you by Ved first. So if you pressed Esc, you stopped
moving the entity, but the entity was still there. And to make matters
worse, it wasn't in an UNRE operation, so you couldn't undo to remove
the duplicate entity, and removing the duplicate entity yourself made
another UNRE operation.

But now this is fixed.