Support improperly-packaged plugin files
By "improperly-packaged", I mean plugin packages whose root is actually
nested one folder deeper than the root of the package, "package" here
meaning either plugin folder or zip. Further nesting and/or other
tomfoolery related to packaging plugins is not accounted for, other than
this (so Lollipop you better not find ANOTHER way to improperly package
Make Me a DSS Souleye...).

I didn't have to mount a mounted folder inside a mounted folder/zip, and
it didn't seem to work anyway when I tried it. All I had to do was
change the plugin path to point to the nested folder.

Ved will still ignore the folder if starts with a dot ('.') or a hash
('#'), and ignores all such folders starting with dots or hashes inside
the package.

Closes #109.