Account for warp lines in warp BG line indicators
If there are warp lines present in a room, the warping background will
not apply. So, Ved will stop drawing blue and yellow warping sidelines,
and switch to the white and red ones, if it sees there's a warp line in
the current room. Furthermore, it will stop drawing the white thin line
coming from the other, non-warping side of a warping border if it sees
that there is a warp line in the room where that warping border comes
from.