Account for hardcoded (19,8) with warping sidelines
(19,8) is hardcoded to warp on all sides no matter what, unless you add
warp lines. Changing the warp background will do nothing for (19,8)'s
warping.

Ved will now assume (19,8) is warping on all sides unless there is a
warp line in it. This means it will show the warping sidelines inside
(19,8), as well as the thin white line for rooms adjacent to (19,8), as
long as it doesn't contain any warp lines.