Add support for drag/dropping files in graphics viewer
This adds the `love.filedropped()` callback, which only works with LÖVE
versions 0.10.0 or above. As well, there is now text saying "Click on
Load \\ or drag and drop onto here" if there is no image displaying in
the graphics viewer, but dragging-and-dropping works even if there is
already an image loaded, and it even works if you drop it anywhere onto
the window and doesn't necessarily have to be on the image canvas.

If the user has a LÖVE version that doesn't have the callback, the text
that appears when no image is loaded will just say "Click on Load"
instead.

I've factored out the image-opening part of
`dialog.callback.openimage()` into its own separate function.