Relax requirements for scrolling fields in dialogs
Previously, you actually had to put your mouse on the scrollable field
in question in order to be able to scroll it, which seemed a bit too
strict because you had to do that in order to be able to use
PageUp/PageDown, for example.

So along with putting your mouse on a field, you're also now allowed to
scroll a field if that field is focused and input is made via keyboard,
or input is made via keyboard and there's only one scrollable field
anyway.